Extract tabs into seperate files
