+++ src/pages/contract/Contract.jsx
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import { Route, Switch, Redirect, Link } from 'react-router-dom';
-import Highlight from 'react-highlight';
+import { Route, Switch, Redirect } from 'react-router-dom';
 import contractStore from '../../store/ContractStore';
 import addressStore from '../../store/AddressStore';
-import uiStore from '../../store/UIStore';
 import RouterUtils from '../../lib/RouterUtils';
 import TextUtils from '../../lib/TextUtils';
 import Loading from '../../components/Loading';
@@ -12,9 +10,9 @@ import HashLink from '../../components/HashLink';
 import ItemNotFound from '../../components/ItemNotFound';
 import AssetsBalancesTable from '../../components/AssetsBalancesTable';
 import PageTitle from '../../components/PageTitle';
-import ItemsTable from '../../components/ItemsTable';
 import Page from '../../components/Page';
-import { Tabs, TabHead, TabBody, Tab, TabPanel } from '../../components/tabs';
+import { Tabs, TabHead, TabBody, Tab } from '../../components/tabs';
+import { AssetsTab, CodeTab, CommandsTab, TransactionsTab } from './components/tabs';
 import './Contract.css';
 
 class ContractPage extends Component {
@@ -109,177 +107,16 @@ class ContractPage extends Component {
         </TabHead>
         <TabBody>
           <Switch>
-            <Route path={`${currentPath}/txns`} component={TransactionsTabReactive} />
-            <Route path={`${currentPath}/commands`} component={CommandsTabReactive} />
+            <Route path={`${currentPath}/txns`} component={TransactionsTab} />
+            <Route path={`${currentPath}/commands`} component={CommandsTab} />
             <Route path={`${currentPath}/code`} component={CodeTab} />
-            <Route path={`${currentPath}/assets`} component={AssetsTabReactive} />
+            <Route path={`${currentPath}/assets`} component={AssetsTab} />
             <Redirect from={`${currentPath}`} to={`${currentPath}/txns`} />
           </Switch>
         </TabBody>
       </Tabs>
     );
   }
-}
-
-const TransactionsTab = observer(props => {
-  return (
-    <TabPanel>
-      <ItemsTable
-        columns={[
-          {
-            Header: 'TX HASH',
-            accessor: 'hash',
-            Cell: data => <HashLink url={`/tx/${data.value}`} hash={data.value} />,
-          },
-          {
-            Header: 'Timestamp',
-            accessor: 'Block.timestamp',
-            Cell: data => TextUtils.getDateStringFromTimestamp(data.value),
-          },
-          {
-            Header: 'Block',
-            accessor: 'Block.blockNumber',
-            Cell: data => <Link to={`/blocks/${data.value}`}>{data.value}</Link>,
-          },
-          {
-            Header: 'First Command',
-            accessor: 'firstCommand',
-          },
-        ]}
-        loading={addressStore.loading.addressTransactions}
-        itemsCount={addressStore.addressTransactionsCount}
-        items={addressStore.addressTransactions}
-        pageSize={uiStore.addressTxsTable.pageSize}
-        curPage={uiStore.addressTxsTable.curPage}
-        tableDataSetter={uiStore.setAddressTxsTableData.bind(uiStore)}
-      />
-    </TabPanel>
-  );
-});
-const TransactionsTabReactive = observer(
-  WithSetAddressOnUiStore(TransactionsTab, 'setAddressTxsTableData')
-);
-
-const CommandsTab = observer(() => {
-  return (
-    <TabPanel>
-      <ItemsTable
-        columns={[
-          {
-            Header: 'Command',
-            accessor: 'command',
-          },
-          {
-            Header: 'Timestamp',
-            accessor: 'Transaction.Block.timestamp',
-            Cell: data => TextUtils.getDateStringFromTimestamp(data.value),
-          },
-          {
-            Header: 'TX Hash',
-            accessor: 'Transaction.hash',
-            Cell: data => <HashLink url={`/tx/${data.value}`} hash={data.value} />,
-          },
-          {
-            Header: 'Block',
-            accessor: 'Transaction.Block.blockNumber',
-            Cell: data => <Link to={`/blocks/${data.value}`}>{data.value}</Link>,
-          },
-        ]}
-        loading={contractStore.loading.commands}
-        itemsCount={contractStore.commandsCount}
-        items={contractStore.commands}
-        pageSize={uiStore.contractCommandsTable.pageSize}
-        curPage={uiStore.contractCommandsTable.curPage}
-        tableDataSetter={uiStore.setContractCommandsTableData.bind(uiStore)}
-      />
-    </TabPanel>
-  );
-});
-const CommandsTabReactive = observer(
-  WithSetAddressOnUiStore(CommandsTab, 'setContractCommandsTableData')
-);
-
-function CodeTab() {
-  if (contractStore.loading.contract) {
-    return <Loading />;
-  }
-  return (
-    <TabPanel>
-      <Highlight className="fsharp">{contractStore.contract.code}</Highlight>
-    </TabPanel>
-  );
-}
-
-const AssetsTab = observer(props => {
-  return (
-    <TabPanel>
-      <ItemsTable
-        columns={[
-          {
-            Header: 'ASSET',
-            accessor: 'asset',
-            Cell: data => <HashLink hash={data.value} />,
-          },
-          {
-            Header: 'TOKENS OUTSTANDING',
-            accessor: 'outstanding',
-            Cell: data => TextUtils.formatNumber(data.value),
-          },
-          {
-            Header: 'TOTAL ISSUED',
-            accessor: 'issued',
-            Cell: data => TextUtils.formatNumber(data.value),
-          },
-          {
-            Header: 'TOKENS DESTROYED',
-            accessor: 'destroyed',
-            Cell: data => TextUtils.formatNumber(data.value),
-          },
-          {
-            Header: 'UNIQUE KEYHOLDERS',
-            accessor: 'keyholders',
-            Cell: data => TextUtils.formatNumber(data.value),
-          },
-        ]}
-        loading={contractStore.loading.assets}
-        itemsCount={contractStore.assetsCount}
-        items={contractStore.assets}
-        pageSize={uiStore.contractAssetsTable.pageSize}
-        curPage={uiStore.contractAssetsTable.curPage}
-        tableDataSetter={uiStore.setContractAssetsTableData.bind(uiStore)}
-      />
-    </TabPanel>
-  );
-});
-const AssetsTabReactive = observer(
-  WithSetAddressOnUiStore(AssetsTab, 'setContractAssetsTableData')
-);
-
-function WithSetAddressOnUiStore(WrappedComponent, uiStoreFunctionName) {
-  return class HOC extends Component {
-    componentDidMount() {
-      this.setAddress();
-    }
-
-    componentDidUpdate(prevProps) {
-      const prevParams = RouterUtils.getRouteParams(prevProps);
-      if (this.addressProp !== prevParams.address) {
-        this.setAddress();
-      }
-    }
-
-    get addressProp() {
-      return RouterUtils.getRouteParams(this.props).address;
-    }
-
-    setAddress() {
-      uiStore[uiStoreFunctionName]({ address: this.addressProp });
-    }
-
-    render() {
-      return <WrappedComponent {...this.props} />;
-    }
-  };
 }
 
 export default observer(ContractPage);
