+++ src/pages/contract/components/WithSetAddressOnUiStore/index.js
@@ -1,0 +1,1 @@
+export { default } from './WithSetAddressOnUiStore.jsx';
