+++ src/pages/contract/components/tabs/CommandsTab.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import { observer } from 'mobx-react';
+import { Link } from 'react-router-dom';
+import contractStore from '../../../../store/ContractStore';
+import uiStore from '../../../../store/UIStore';
+import TextUtils from '../../../../lib/TextUtils';
+import WithSetAddressOnUiStore from '../WithSetAddressOnUiStore';
+import { TabPanel } from '../../../../components/tabs';
+import ItemsTable from '../../../../components/ItemsTable';
+import HashLink from '../../../../components/HashLink';
+
+const CommandsTab = observer(() => {
+  return (
+    <TabPanel>
+      <ItemsTable
+        columns={[
+          {
+            Header: 'Command',
+            accessor: 'command',
+          },
+          {
+            Header: 'Timestamp',
+            accessor: 'Transaction.Block.timestamp',
+            Cell: data => TextUtils.getDateStringFromTimestamp(data.value),
+          },
+          {
+            Header: 'TX Hash',
+            accessor: 'Transaction.hash',
+            Cell: data => <HashLink url={`/tx/${data.value}`} hash={data.value} />,
+          },
+          {
+            Header: 'Block',
+            accessor: 'Transaction.Block.blockNumber',
+            Cell: data => <Link to={`/blocks/${data.value}`}>{data.value}</Link>,
+          },
+        ]}
+        loading={contractStore.loading.commands}
+        itemsCount={contractStore.commandsCount}
+        items={contractStore.commands}
+        pageSize={uiStore.contractCommandsTable.pageSize}
+        curPage={uiStore.contractCommandsTable.curPage}
+        tableDataSetter={uiStore.setContractCommandsTableData.bind(uiStore)}
+      />
+    </TabPanel>
+  );
+});
+export default observer(WithSetAddressOnUiStore(CommandsTab, 'setContractCommandsTableData'));
