+++ src/pages/contract/components/tabs/index.js
@@ -1,0 +1,4 @@
+export { default as AssetsTab } from './AssetsTab.jsx';
+export { default as CodeTab } from './CodeTab.jsx';
+export { default as CommandsTab } from './CommandsTab.jsx';
+export { default as TransactionsTab } from './TransactionsTab.jsx';
